# public/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: 'Takebook API'
  contact:
    email: support@example.com
  version: 1.0.0
paths:
...
  /users/me:
    get:
      tags:
        - users
      summary: 'Retorna o usuário logado'
      operationId: GetAuthUser
      responses:
        '200':
          description: ...
      security:
        -
          apiToken: []
    put:
      tags:
        - users
      summary: 'Atualiza o usuário'
      operationId: UpdateAuthUser
      parameters:
        -
          name: first_name
          in: query
          description: 'Primeiro nome'
          required: false
          schema:
            type: string
        -
          name: last_name
          in: query
          description: 'Ultimo nome'
          required: false
          schema:
            type: string
        -
          name: address_street
          in: query
          description: 'Rua onde o rapaz mora'
          required: false
          schema:
            type: string
        -
          name: address_number
          in: query
          description: 'Numero da casa do rapaz'
          required: false
          schema:
            type: string
        -
          name: address_neighborhood
          in: query
          description: 'Bairro onde o rapaz mora'
          required: false
          schema:
            type: string
        -
          name: address_city
          in: query
          description: 'Cidade onde o rapaz mora'
          required: false
          schema:
            type: string
        -
          name: address_state
          in: query
          description: 'Rua onde o rapaz mora'
          required: false
          schema:
            type: string
        -
          name: address_zip_code
          in: query
          description: 'CEP do cidadão'
          required: false
          schema:
            type: string
        -
          name: email
          in: query
          description: 'E-mail do cidadão'
          required: false
          schema:
            type: string
      responses:
        '200':
          description: ...
      security:
        -
          apiToken: []
  /users/me/change:
    put:
      tags:
        - users
      summary: 'Atualiza a senha do usuário'
      operationId: UpdateAuthUserPassword
      parameters:
        -
          name: old_password
          in: query
          description: 'Senha antiga do usuário'
          required: true
          schema:
            type: string
        -
          name: password
          in: query
          description: 'Nova senha do usuário'
          required: true
          schema:
            type: string
        -
          name: password_confirmation
          in: query
          description: 'Confirmação da senha'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ...
      security:
        -
          apiToken: []
  /users/me/likes:
    get:
      tags:
        - users
      summary: 'Retorna todos os livros likados do usuário logado'
      operationId: UserLikedBooks
      responses:
        '200':
          description: ...
      security:
        -
          apiToken: []
  '/users/me/likes/{book_id}':
    post:
      tags:
        - users
      summary: 'Dá um like em um livro para o usuário logado'
      operationId: UserLikeBook
      parameters:
        -
          name: book_id
          in: path
          description: 'ID do livro'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ...
      security:
        -
          apiToken: []
...
components:
  securitySchemes:
    apiToken:
      type: apiKey
      description: 'Token do client para consumo da API'
      name: Authorization
      in: header
      scheme: bearer